Separated FullpageSectionController.svelte from FullpageSection.svelte
